Context, User: Ensure log-in page is not the first loaded page in the session.
We believe this to be a cause of form field manipulation intermittently
failing, causing the log-in step to fail -- as "Given I am logged in as a user"
is often the the first line in a scenario!

See https://github.com/Behat/MinkExtension/issues/286

This is the latest attempt in a long line of fixes to make Behat user log-in
reliable. See #47 for the history.